Split Agent 365 reactive and proactive examples
