+++ examples/agent365/src/main.py
@@ -3,41 +3,40 @@ Copyright (c) Microsoft Corporation. All rights reserved.
 Licensed under the MIT License.
 """
 
-import argparse
 import asyncio
 import logging
+import re
 
-from microsoft_teams.api import MessageActivityInput
-from microsoft_teams.apps import App
+from microsoft_teams.api import MessageActivity
+from microsoft_teams.api.activities.typing import TypingActivityInput
+from microsoft_teams.apps import ActivityContext, App
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
+app = App()
 
-async def main():
-    parser = argparse.ArgumentParser(description="Send a proactive message using API-level AgenticIdentity")
-    parser.add_argument("conversation_id", help="The Teams conversation ID to send messages to")
-    parser.add_argument("agentic_app_id", help="The concrete agent identity app/client ID")
-    parser.add_argument("agentic_user_id", help="The agent user object ID")
-    args = parser.parse_args()
 
-    app = App()
-    await app.initialize()
+@app.on_message_pattern(re.compile(r"hello|hi|greetings"))
+async def handle_greeting(ctx: ActivityContext[MessageActivity]) -> None:
+    """Handle greeting messages using the inbound AgenticIdentity when present."""
+    await ctx.reply("Hello! How can I assist you today?")
 
-    agentic_identity = app.get_agentic_identity(args.agentic_app_id, args.agentic_user_id)
-    sent = await app.send(
-        args.conversation_id,
-        "Hello from app.send with an AgenticIdentity.",
-        agentic_identity=agentic_identity,
-    )
-    logger.info("Sent activity through app.send. Activity ID: %s", sent.id)
 
-    api_sent = await app.api.conversations.activities(args.conversation_id).create(
-        MessageActivityInput(text="Hello from the conversation activity API with an AgenticIdentity."),
-        agentic_identity=agentic_identity,
-    )
-    logger.info("Sent activity through app.api. Activity ID: %s", api_sent.id)
+@app.on_message
+async def handle_message(ctx: ActivityContext[MessageActivity]):
+    """Echo incoming messages using the inbound AgenticIdentity when present."""
+    logger.info("[Agent365 reactive] Message received: %s", ctx.activity.text)
+    logger.info("[Agent365 reactive] From: %s", ctx.activity.from_)
+    logger.info("[Agent365 reactive] Agentic identity: %s", ctx.activity.recipient.agentic_identity)
+
+    await ctx.reply(TypingActivityInput())
+
+    if "reply" in ctx.activity.text.lower():
+        await ctx.reply("Hello! How can I assist you today?")
+    else:
+        await ctx.send(f"You said '{ctx.activity.text}'")
 
 
 if __name__ == "__main__":
-    asyncio.run(main())
+    asyncio.run(app.start())
